fix(probability): bound replay candidates by producer limits
